# Optimizations for Armv8-A