This commit adapts the past SIMD optimizations to the Neon extension that is part of the Armv8-A architecture. The changes apply only to the AArch64 execution state (32-bit code would require more work due to the smaller general-purpose registers, so I didn't bother).

In order to gather some performance data, I compiled the included benchmark `bench.c` using:
```gcc -Wall -Wextra -Ofast -flto -march=native -g -o bench bench.c picohttpparser.c```
And then ran it with:
```taskset -c 1 time -f "%e" ./bench```

I ran the benchmark on Ubuntu 18.04 using 3 different cloud instances:

- [`a1.large`](https://aws.amazon.com/ec2/instance-types/a1/) on Amazon Web Services - uses AWS Graviton processors that are apparently based on Arm Cortex-A72
- [`c1.large.arm`](https://www.packet.com/cloud/servers/c1-large-arm) on Packet - uses Cavium ThunderX processors (note that it is the first version)
- [`c2.large.arm`](https://www.packet.com/cloud/servers/c2-large-arm) also on Packet - uses Ampere eMAG processors

Here are the median results from 20 runs - all times are in seconds:

Instance|Base time|New time|Change
-|-|-|-
a1.large|7.62|4.63|-39.24%
c1.large.arm|16.23|13.36|-17.68%
c2.large.arm|6.29|6.88|+9.38%

Standard errors are 0.1% or less in all cases.

I don't have a good explanation for the regression on Ampere eMAG right now, but I noticed that compiling with Clang produced slightly better times (though still slower than the baseline), so a probable partial explanation is that the software support for the microarchitecture (which is the most recent one) can be improved (or it will be a while until the enhancements make their way into the OS images that can actually be deployed). Unfortunately, I couldn't find any optimization guide for the processor, and the support for the hardware performance counters seemed flaky, so it was a bit difficult to do a deeper analysis.

It should also be possible to optimize the `parse_headers()` function using the `TBL` instruction, but that would require transforming `token_char_map` into a bit array (or something similar), so that it fits into at most 4 vector registers.

I also have an initial implementation (not tested much and certainly not benchmarked) using the Scalable Vector Extension (SVE) in a [branch](https://github.com/volyrique/picohttpparser/tree/armv8-a+sve) in my fork of the repository.