@kazuho 2019-05-17
Thank you for the PR.

The results are interesting. I'll check the numbers on A52.