@volyrique 2019-06-02
I figured out that header values had a higher chance of being large, so I decided to unroll the vector loop in `get_token_to_eol()` a bit. Here are the new results:

Instance|Base time|New time|Change
-|-|-|-
a1.large|7.62|5.13|-32.68%
c1.large.arm|16.23|12.54|-22.74%
c2.large.arm|6.29|6.20|-1.43%

The result on the Amazon instance is a bit worse (but still significantly faster than the scalar version), while all the other values have improved; in particular, there is no longer a performance regression on the Ampere eMAG. Standard errors are 0.26% or less.