@volyrique 2019-10-31
Now that Travis CI [supports](https://blog.travis-ci.com/2019-10-07-multi-cpu-architecture-support) testing in an Arm64 environment, I have enabled it for this project.

I think I also have a pretty good idea about why the performance on the Ampere eMAG is not that good. After some experiments, I have determined that the vector instruction throughput on that machine is 0.50 (instructions per cycle), while on Arm Cortex-A72 it is 1.49 (probably 1.5 - there is some measurement noise). Those values are for vector bitwise operations and comparisons, which are the main operations executed by my optimization. As a result, it is worth vectorizing on the Ampere machine mainly if there is a significant amount of data to process, so it is not surprising that the second version of my changes, which has raised the threshold for switching from scalar to vector code, behaves better.

As for the hardware performance counters being problematic on the Ampere eMAG - it turns out that there are no problems if the counters are specified explicitly on the `perf` command line using the IDs from the [Ampere documentation](https://developer.amperecomputing.com/resources/).