@volyrique 2023-07-20
Optimized implementations of the `parse_token()` function using the `TBL` instruction or SVE gather loads are available in [this](https://github.com/volyrique/picohttpparser/tree/armv8-a+tbl) and [that](https://github.com/volyrique/picohttpparser/tree/armv8-a+sve) branch respectively, but neither of them are as convincing as the changes I have proposed here.

The regular SVE optimizations have been merged into this PR.